fix migrations and seeders

+++ database/migrations/2022_08_11_140430_create_drivers_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
     {
         Schema::create('drivers', function (Blueprint $table) {
             $table->id();
-            $table->integer('identification_number')->length(10);
+            $table->bigInteger('identification_number')->length(10);
             $table->string('birthdate', 10);
             $table->string('name');
             $table->string('driving_percentage');
