+++ database/migrations/2022_08_11_140432_create_cars_table.php
@@ -15,9 +15,9 @@ return new class extends Migration
     {
         Schema::create('cars', function (Blueprint $table) {
             $table->id();
-            $table->integer('customs_serial_number')->length(10)->nullable();
-            $table->integer('customs_form_number')->length(10)->nullable();
-            $table->integer('car_id')->length(20);
+            $table->bigInteger('customs_serial_number')->length(10)->nullable();
+            $table->bigInteger('customs_form_number')->length(10)->nullable();
+            $table->bigInteger('car_id')->length(20);
             $table->integer('car_plate_code')->length(2);
             $table->integer('car_plate_number')->length(5);
             $table->string('production_year');
